+++ dbus_next/__init__.py
@@ -1,5 +1,3 @@
-from . import aio
-from . import glib
 from .constants import (BusType, MessageType, MessageFlag, NameFlag, RequestNameReply,
                         ReleaseNameReply, PropertyAccess, ArgDirection, ErrorType)
 from .errors import (SignatureBodyMismatchError, InvalidSignatureError, InvalidAddressError,
@@ -9,8 +7,8 @@ from .errors import (SignatureBodyMismatchError, InvalidSignatureError, InvalidA
                      DBusError)
 from . import introspection
 from .message import Message
-from . import message_bus
-from . import proxy_object
+from .message_bus import MessageBus, ValueEvent
+from .proxy_object import ProxyObject, ProxyInterface
 from . import service
 from .signature import SignatureType, SignatureTree, Variant
 from .validators import (is_bus_name_valid, is_object_path_valid, is_interface_name_valid,