+++ examples/aio-list-names.py
@@ -3,8 +3,8 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.dirname(__file__) + '/..'))
 
-from dbus_next import Message, MessageType
-from dbus_next.aio import MessageBus
+from asyncdbus import Message, MessageType
+from asyncdbus import MessageBus
 
 import anyio
 import json