+++ examples/aio-tcp-notification.py
@@ -12,7 +12,7 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.dirname(__file__) + '/..'))
 
-from dbus_next.aio import MessageBus
+from asyncdbus import MessageBus
 
 import anyio