+++ test/client/test_properties.py
@@ -1,11 +1,7 @@
-from asyncdbus import aio, glib, Message, DBusError
+from asyncdbus import Message, DBusError, MessageBus
 from asyncdbus.service import ServiceInterface, dbus_property, PropertyAccess
-from test.util import check_gi_repository, skip_reason_no_gi
 
 import pytest
-
-has_gi = check_gi_repository()
-
 
 class ExampleInterface(ServiceInterface):
     def __init__(self):
@@ -38,11 +34,11 @@ class ExampleInterface(ServiceInterface):
 
 @pytest.mark.anyio
 async def test_aio_properties():
-  async with aio.MessageBus().connect() as service_bus:
+  async with MessageBus().connect() as service_bus:
    service_interface = ExampleInterface()
    service_bus.export('/test/path', service_interface)
 
-   async with aio.MessageBus().connect() as bus:
+   async with MessageBus().connect() as bus:
     obj = bus.get_proxy_object(service_bus.unique_name, '/test/path',
                                service_bus._introspect_export_path('/test/path'))
     interface = obj.get_interface(service_interface.name)
@@ -76,40 +72,3 @@ async def test_aio_properties():
             raise e
 
 
-@pytest.mark.skipif(not has_gi, reason=skip_reason_no_gi)
-def test_glib_properties():
-    service_bus = glib.MessageBus().connect_sync()
-    service_interface = ExampleInterface()
-    service_bus.export('/test/path', service_interface)
-
-    bus = glib.MessageBus().connect_sync()
-    obj = bus.get_proxy_object(service_bus.unique_name, '/test/path',
-                               service_bus._introspect_export_path('/test/path'))
-    interface = obj.get_interface(service_interface.name)
-
-    prop = interface.get_some_property_sync()
-    assert prop == service_interface._some_property
-
-    interface.set_some_property_sync('different')
-    assert service_interface._some_property == 'different'
-
-    with pytest.raises(DBusError):
-        try:
-            prop = interface.get_error_throwing_property_sync()
-            assert False, prop
-        except DBusError as e:
-            assert e.type == service_interface.error_name
-            assert e.text == service_interface.error_text
-            assert type(e.reply) is Message
-            raise e
-
-    with pytest.raises(DBusError):
-        try:
-            interface.set_error_throwing_property_sync('different2')
-        except DBusError as e:
-            assert e.type == service_interface.error_name
-            assert e.text == service_interface.error_text
-            assert type(e.reply) is Message
-            raise e
-
-    service_bus.disconnect()