+++ test/test_big_message.py
@@ -1,11 +1,7 @@
-from asyncdbus import aio, glib, Message, MessageType
+from asyncdbus import MessageBus, Message, MessageType
 from asyncdbus.service import ServiceInterface, method
-from test.util import check_gi_repository, skip_reason_no_gi
 
 import pytest
-
-has_gi = check_gi_repository()
-
 
 class ExampleInterface(ServiceInterface):
     def __init__(self):
@@ -18,9 +14,9 @@ class ExampleInterface(ServiceInterface):
 
 @pytest.mark.anyio
 async def test_aio_big_message():
-  'this tests that nonblocking reads and writes actually work for aio'
-  async with aio.MessageBus().connect() as bus1, \
-          aio.MessageBus().connect() as bus2:
+  'this tests that nonblocking reads and writes actually work'
+  async with MessageBus().connect() as bus1, \
+          MessageBus().connect() as bus2:
     interface = ExampleInterface()
     bus1.export('/test/path', interface)
 
@@ -36,23 +32,3 @@ async def test_aio_big_message():
     assert result.message_type == MessageType.METHOD_RETURN, result.body[0]
     assert result.body[0] == big_body[0]
 
-
-@pytest.mark.skipif(not has_gi, reason=skip_reason_no_gi)
-def test_glib_big_message():
-    'this tests that nonblocking reads and writes actually work for glib'
-    bus1 = glib.MessageBus().connect_sync()
-    bus2 = glib.MessageBus().connect_sync()
-    interface = ExampleInterface()
-    bus1.export('/test/path', interface)
-
-    # two megabytes
-    big_body = [bytes(1000000) * 2]
-    result = bus2.call_sync(
-        Message(destination=bus1.unique_name,
-                path='/test/path',
-                interface=interface.name,
-                member='echo_bytes',
-                signature='ay',
-                body=big_body))
-    assert result.message_type == MessageType.METHOD_RETURN, result.body[0]
-    assert result.body[0] == big_body[0]